requirements.txt
alembic==1.13.1
async-timeout==4.0.3
blinker==1.8.2
cachelib==0.9.0
click==8.1.7
exceptiongroup==1.2.1
Flask==3.0.3
Flask-Caching==2.3.0
Flask-Migrate==4.0.7
Flask-Script==2.0.6
Flask-SQLAlchemy==3.1.1
gevent==24.2.1
greenlet==3.0.3
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==2.1.5
packaging==24.1
pluggy==1.5.0
pytest==8.2.2
pytest-mock==3.14.0
python-dotenv==1.0.1
redis==5.0.6
SQLAlchemy==2.0.30
tomli==2.0.1
typing_extensions==4.12.2
Werkzeug==3.0.3
zope.event==5.0
zope.interface==6.4.post2
